+++ src/functions/init/intro.js
@@ -15,24 +15,28 @@ const ConfigManager = require("../../classes/ConfigManager");
 const InfoMessages = require("../../classes/InfoMessages");
 
 const config = new ConfigManager();
+const configData = config.getConfig();
 
 // Detection if configuration file is corrupt
 // TODO make this into a seperate function due to usage in 'history' command
-if (typeof config.getConfig() === "undefined") {
+if (typeof configData === "undefined") {
   // InfoMessages.error("Error when reading the configuration file. Resetting file...");
 
   config.deleteConfig();
   config.createConfig();
 }
 
-console.log(`${chalk.bold(`${GLOBAL_NAME}, v${VERSION} (build ${BUILD})`)}`);
+if (configData.settings?.showVersionOnStart?.current?.value) {
+  console.log(`${chalk.bold(`${GLOBAL_NAME}, v${VERSION} (build ${BUILD})`)}`);
+}
 
 // Only display if this is the first time BubbleOS is runing
-if (!config.getConfig().firstIntro) {
+if (!configData.firstIntro) {
   console.log(`Made by ${AUTHOR}!`);
 }
 
-console.log();
+if (configData.settings?.showVersionOnStart?.current?.value || !configData.firstIntro)
+  console.log();
 
 Verbose.custom(`Checking if ${GLOBAL_NAME} is in beta...`);
 if (IN_BETA) {
@@ -63,14 +67,14 @@ if (IN_BETA) {
 }
 
 // Show information about commands
-if (!config.getConfig().firstIntro) {
+if (!configData.firstIntro) {
   console.log(`For a list on some available commands, type ${chalk.italic("'help'")}.`);
   console.log(`For more information about a command, type ${chalk.italic("'help <command>'")}.\n`);
 
   console.log(`To exit the ${GLOBAL_NAME} shell, type ${chalk.italic("'exit'")}.\n`);
 }
 
-if (config.getConfig().lastCrashed) {
+if (configData.lastCrashed) {
   InfoMessages.warning(
     `${GLOBAL_NAME} crashed the last time it was run. For more information, find the error information file in the directory you ran ${GLOBAL_NAME} in.`
   );