+++ src/variables/commands.js
@@ -25,6 +25,7 @@ const ping = require("../commands/ping");
 const print = require("../commands/print");
 const readfile = require("../commands/readfile");
 const rename = require("../commands/rename");
+const setman = require("../commands/setman");
 const stat = require("../commands/stat");
 const symlink = require("../commands/symlink");
 const sysinfo = require("../commands/sysinfo");
@@ -79,6 +80,7 @@ const COMMANDS = {
   echo: print,
   readfile,
   rename,
+  setman,
   stat,
   symlink,
   sysinfo,
@@ -102,7 +104,7 @@ const ALIASES = {
   cls: [],
   copy: ["cp", "copyfile", "copydir"],
   crash: ["destroy"],
-  cwd: ["pwd"],
+  cwd: [],
   date: [],
   del: ["rm", "delete", "rmfile", "rmdir"],
   dirtree: [],
@@ -121,6 +123,7 @@ const ALIASES = {
   print: [],
   readfile: ["cat", "more", "type", "rdfile", "read", "tail"],
   rename: ["mv", "ren"],
+  setman: ["settings", "syscfg", "sysconfig", "sysedit"],
   stat: ["df", "fileinfo", "dirinfo", "pathinfo", "size"],
   symlink: ["symblnk", "ln", "link"],
   sysinfo: ["uname", "systeminfo", "userinfo", "whoami"],